extract tile_model
